+++ src/components/FallbackMode/index.ts
@@ -1,0 +1,1 @@
+export { FallbackMode } from './FallbackMode'